support for more than 32 FN keys

+++ tmk_core/common/keymap.h
@@ -42,10 +42,10 @@ typedef union {
 
 
 /* translates key to keycode */
-uint8_t keymap_key_to_keycode(uint8_t layer, keypos_t key);
+uint16_t keymap_key_to_keycode(uint8_t layer, keypos_t key);
 
 /* translates Fn keycode to action */
-action_t keymap_fn_to_action(uint8_t keycode);
+action_t keymap_fn_to_action(uint16_t keycode);
